requirements.txt
mlflow
opencv-python
Pillow
Django
tensorflow-cpu
openai
django-cors-headers
djangorestframework-simplejwt
scikit-surprise